Refactored server code into single folder
Refactored routes to routes folder
Merged codebases into single server folder
Returned listen to port 8080
Renamed some endpoints to read more like plain English, e.g. /createUser is now /create/newUser
