--- index.html
@@ -1,1 +1,0 @@
-<div>VSCode with GitHub Practice</div>
