+++ server/package.json
@@ -1,25 +1,38 @@
 {
-  "name": "server",
+  "name": "cse-media",
   "version": "1.0.0",
-  "description": "",
+  "description": "This GitHub repo contains the project and project artifacts for the Lab section of CSE 6214.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.1.0",
+    "body-parser": "^1.20.0",
+    "bootstrap": "^5.2.2",
+    "chai-http": "^4.3.0",
     "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.1",
     "express-session": "^1.17.3",
     "helmet": "^6.0.0",
+    "nodemon": "^2.0.20",
     "pg": "^8.8.0",
+    "postgres-migrations": "^5.3.0",
+    "react-bootstrap": "^2.5.0",
+    "react-query": "^3.39.2",
+    "react-router-dom": "^6.4.2",
     "session": "^0.1.0",
     "sessions": "^0.0.2-7",
     "socket.io": "^4.5.2",
     "yup": "^0.32.11"
+  },
+  "devDependencies": {
+    "chai": "^4.3.6",
+    "mocha": "^10.0.0"
   }
 }
